+++ lib/shop_page.dart
@@ -405,7 +405,7 @@ class _ShopPageState extends State<ShopPage> {
 
   bool _handleScrollNotification(ScrollNotification notification) {
     if (notification is ScrollUpdateNotification) {
-      final currentOffset = max(notification.metrics.pixels, 0);
+      final currentOffset = max(notification.metrics.pixels, 0.0).toDouble();
 
       if (currentOffset <= 0) {
         _lastScrollOffset = 0;
